requirements.txt
asgiref==3.5.2
Django==4.1.3
django-base64field==1.0
django-cors-headers==3.13.0
django-restapi==0.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
face-recognition-models==0.3.0
psycopg2==2.9.5
PyJWT==1.7.1
pytz==2022.6
sqlparse==0.4.3
